Moved IT test to own package
